api/requirements.txt
asgiref==3.5.2
Django==4.1.3
django-cors-headers==3.13.0
django-debug-toolbar==3.7.0
django-environ==0.9.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.6.0
pytz==2022.6
sqlparse==0.4.3
